+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Improved integer/float detection in `xlsx_raw` using modulo operator
+* Improved test coverage and structure
 
 ### Fixed